Refactor of VerbalFlashcardsPage.js, with seperate components folder
